requirements.txt
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
rich==13.4.2 